requirements.txt
beautifulsoup4==4.12.2
boto3==1.28.68
botocore==1.31.68
certifi==2023.7.22
charset-normalizer==3.3.0
cramjam==2.7.0
fastparquet==2023.8.0
fsspec==2023.10.0
idna==3.4
jmespath==1.0.1
numpy==1.26.1
packaging==23.2
pandas==2.1.1
pyarrow==13.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.0.7
